Move CLI to conformance package
